refactor(frontend): rename ic types to ic-tokens
