+++ src/frontend/src/icp-eth/derived/icrc-erc20.derived.ts
@@ -2,7 +2,7 @@ import { enabledErc20TokensAddresses } from '$eth/derived/erc20.derived';
 import type { ContractAddressText } from '$eth/types/address';
 import type { Erc20ContractAddress, Erc20Token } from '$eth/types/erc20';
 import { enabledIcrcTokens } from '$icp/derived/icrc.derived';
-import type { IcCkToken, IcToken } from '$icp/types/ic';
+import type { IcCkToken, IcToken } from '$icp/types/ic-tokens';
 import { nonNullish } from '@dfinity/utils';
 import { derived, type Readable } from 'svelte/store';
