+++ src/frontend/src/icp/components/convert/ConvertToBTC.svelte
@@ -5,7 +5,7 @@
 	import { BTC_MAINNET_SYMBOL } from '$env/tokens.btc.env';
 	import IcSendModal from '$icp/components/send/IcSendModal.svelte';
 	import { ckBtcMinterInfoStore } from '$icp/stores/ckbtc.store';
-	import type { OptionIcCkToken } from '$icp/types/ic';
+	import type { OptionIcCkToken } from '$icp/types/ic-tokens';
 	import ButtonHero from '$lib/components/hero/ButtonHero.svelte';
 	import IconCkConvert from '$lib/components/icons/IconCkConvert.svelte';
 	import { isBusy } from '$lib/derived/busy.derived';
