+++ src/frontend/src/icp/components/fee/IcTokenFee.svelte
@@ -2,7 +2,7 @@
 	import { nonNullish } from '@dfinity/utils';
 	import { BigNumber } from '@ethersproject/bignumber';
 	import { getContext } from 'svelte';
-	import type { OptionIcToken } from '$icp/types/ic';
+	import type { OptionIcToken } from '$icp/types/ic-tokens';
 	import Value from '$lib/components/ui/Value.svelte';
 	import { i18n } from '$lib/stores/i18n.store';
 	import { SEND_CONTEXT_KEY, type SendContext } from '$lib/stores/send.store';
