+++ src/frontend/src/icp/components/transactions/IcTransactionsCkEthereumListeners.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
 	import IcCkListener from '$icp/components/core/IcCkListener.svelte';
 	import { initCkETHMinterInfoWorker } from '$icp/services/worker.ck-minter-info.services';
-	import type { OptionIcCkToken } from '$icp/types/ic';
+	import type { OptionIcCkToken } from '$icp/types/ic-tokens';
 	import CkEthereumPendingTransactionsListener from '$icp-eth/components/core/CkEthereumPendingTransactionsListener.svelte';
 	import { ckEthereumNativeToken } from '$icp-eth/derived/cketh.derived';
 	import { token } from '$lib/stores/token.store';
