+++ src/frontend/src/icp/derived/ckbtc-transactions.derived.ts
@@ -1,7 +1,7 @@
 import { ckBtcPendingUtxosStore } from '$icp/stores/ckbtc-utxos.store';
 import { ckBtcMinterInfoStore } from '$icp/stores/ckbtc.store';
 import type { IcTransactionsData } from '$icp/stores/ic-transactions.store';
-import type { IcCkToken } from '$icp/types/ic';
+import type { IcCkToken } from '$icp/types/ic-tokens';
 import { mapCkBTCPendingUtxo } from '$icp/utils/ckbtc-transactions.utils';
 import { isTokenCkBtcLedger } from '$icp/utils/ic-send.utils';
 import { tokenWithFallback } from '$lib/derived/token.derived';
