+++ src/frontend/src/icp/services/ckbtc.services.ts
@@ -7,7 +7,7 @@ import { btcAddressStore } from '$icp/stores/btc.store';
 import { ckBtcPendingUtxosStore } from '$icp/stores/ckbtc-utxos.store';
 import { ckBtcMinterInfoStore } from '$icp/stores/ckbtc.store';
 import type { CkBtcUpdateBalanceParams } from '$icp/types/ckbtc';
-import type { IcCkMetadata, IcCkToken, IcToken } from '$icp/types/ic';
+import type { IcCkMetadata, IcCkToken, IcToken } from '$icp/types/ic-tokens';
 import { queryAndUpdate, type QueryAndUpdateRequestParams } from '$lib/actors/query.ic';
 import { ProgressStepsUpdateBalanceCkBtc } from '$lib/enums/progress-steps';
 import { waitWalletReady } from '$lib/services/actions.services';
