+++ src/frontend/src/icp/services/worker.icrc-wallet.services.ts
@@ -3,7 +3,7 @@ import {
 	onLoadTransactionsError,
 	onTransactionsCleanUp
 } from '$icp/services/ic-transactions.services';
-import type { IcToken } from '$icp/types/ic';
+import type { IcToken } from '$icp/types/ic-tokens';
 import type { WalletWorker } from '$lib/types/listener';
 import type {
 	PostMessage,
