+++ src/frontend/src/icp/stores/icrc-default-tokens.store.ts
@@ -1,4 +1,4 @@
 import { initCertifiedIcrcTokensStore } from '$icp/stores/certified-icrc.store';
-import type { IcToken, IcTokenWithoutId } from '$icp/types/ic';
+import type { IcToken, IcTokenWithoutId } from '$icp/types/ic-tokens';
 
 export const icrcDefaultTokensStore = initCertifiedIcrcTokensStore<IcTokenWithoutId, IcToken>();
