+++ src/frontend/src/icp/stores/receive-token.store.ts
@@ -1,7 +1,7 @@
 import { ETHEREUM_NETWORK } from '$env/networks.env';
 import { ETHEREUM_TOKEN } from '$env/tokens.env';
 import type { EthereumNetwork } from '$eth/types/network';
-import type { IcCkToken, IcToken } from '$icp/types/ic';
+import type { IcCkToken, IcToken } from '$icp/types/ic-tokens';
 import type { Token, TokenId, TokenStandard } from '$lib/types/token';
 import { derived, writable, type Readable } from 'svelte/store';
