+++ src/frontend/src/icp/types/ic-token-toggleable.ts
@@ -1,4 +1,4 @@
-import type { IcToken } from '$icp/types/ic';
+import type { IcToken } from '$icp/types/ic-tokens';
 import type { TokenToggleable } from '$lib/types/token-toggleable';
 
 export type IcTokenToggleable = TokenToggleable<IcToken>;
