+++ src/frontend/src/icp/types/icrc-custom-token.ts
@@ -1,5 +1,5 @@
 import type { EnvIcrcToken, EnvIcrcTokenMetadata } from '$env/types/env-icrc-token';
-import type { IcToken, IcTokenWithoutId } from '$icp/types/ic';
+import type { IcToken, IcTokenWithoutId } from '$icp/types/ic-tokens';
 import type { TokenToggleable, UserTokenState } from '$lib/types/token-toggleable';
 import type { Option } from '$lib/types/utils';
