+++ src/frontend/src/icp/utils/ic-transactions.utils.ts
@@ -1,7 +1,7 @@
 import { ICP_TOKEN_ID } from '$env/tokens.env';
 import type { BtcStatusesData } from '$icp/stores/btc.store';
 import type { IcCertifiedTransaction } from '$icp/stores/ic-transactions.store';
-import type { IcCkToken, IcToken } from '$icp/types/ic';
+import type { IcCkToken, IcToken } from '$icp/types/ic-tokens';
 import type {
 	IcTransaction,
 	IcTransactionUi,
