+++ src/frontend/src/icp/utils/icrc-ledger.utils.ts
@@ -1,5 +1,5 @@
 import { ICRC_LEDGER_CANISTER_TESTNET_IDS } from '$env/networks.icrc.env';
-import type { IcToken } from '$icp/types/ic';
+import type { IcToken } from '$icp/types/ic-tokens';
 import { nonNullish } from '@dfinity/utils';
 
 export const isTokenIcrcTestnet = ({ ledgerCanisterId }: Partial<IcToken>): boolean =>
