+++ src/frontend/src/icp/utils/icrc.utils.ts
@@ -1,6 +1,6 @@
 import { ICP_NETWORK, ICP_NETWORK_ID } from '$env/networks.env';
 import type { LedgerCanisterIdText } from '$icp/types/canister';
-import type { IcCkInterface, IcCkToken, IcFee, IcInterface, IcToken } from '$icp/types/ic';
+import type { IcCkInterface, IcCkToken, IcFee, IcInterface, IcToken } from '$icp/types/ic-tokens';
 import type { IcTokenWithoutIdExtended, IcrcCustomToken } from '$icp/types/icrc-custom-token';
 import type { CanisterIdText } from '$lib/types/canister';
 import type { Token, TokenCategory, TokenMetadata } from '$lib/types/token';
