+++ src/frontend/src/icp/utils/wallet.utils.ts
@@ -1,6 +1,6 @@
 import { initIcpWalletWorker } from '$icp/services/worker.icp-wallet.services';
 import { initIcrcWalletWorker } from '$icp/services/worker.icrc-wallet.services';
-import type { IcToken } from '$icp/types/ic';
+import type { IcToken } from '$icp/types/ic-tokens';
 import type { InitWalletWorkerFn } from '$lib/types/listener';
 
 export const initWalletWorker: InitWalletWorkerFn = ({ token }) =>
