+++ src/frontend/src/lib/services/token.services.ts
@@ -1,7 +1,7 @@
 import type { SaveUserToken } from '$eth/services/erc20-user-tokens-services';
 import type { Erc20Token } from '$eth/types/erc20';
 import type { SaveCustomToken } from '$icp/services/ic-custom-tokens.services';
-import type { IcCkToken } from '$icp/types/ic';
+import type { IcCkToken } from '$icp/types/ic-tokens';
 import { busy } from '$lib/stores/busy.store';
 import { toastsError } from '$lib/stores/toasts.store';
 import { token as tokenStore } from '$lib/stores/token.store';
