+++ src/frontend/src/lib/types/post-message.ts
@@ -8,7 +8,7 @@ import type { BitcoinNetwork as SignerBitcoinNetwork } from '$declarations/signe
 import type { BtcAddressData } from '$icp/stores/btc.store';
 import type { JsonText } from '$icp/types/btc.post-message';
 import type { LedgerCanisterIdText } from '$icp/types/canister';
-import type { IcCanisters, IcCkMetadata } from '$icp/types/ic';
+import type { IcCanisters, IcCkMetadata } from '$icp/types/ic-tokens';
 import type { BtcAddress } from '$lib/types/address';
 import type { Network } from '$lib/types/network';
 import type { CertifiedData } from '$lib/types/store';
