+++ src/frontend/src/lib/utils/token-group.utils.ts
@@ -1,4 +1,4 @@
-import type { IcCkToken } from '$icp/types/ic';
+import type { IcCkToken } from '$icp/types/ic-tokens';
 import { isIcCkToken } from '$icp/utils/icrc.utils';
 import { ZERO } from '$lib/constants/app.constants';
 import type { TokenUi } from '$lib/types/token';
